requirements.txt
pyarrow==14.0.1
apache-airflow[amazon]==3.0.6
apache-airflow-providers-slack==9.3.0
pandas==2.3.3
requests==2.32.5
pydantic-settings==2.11.0
awscli==1.42.59
boto3==1.40.59
oauth2client==4.1.3
gspread===6.2.1






